Move new function to cleaning utils
