+++ jobs/validate_cleaned_ind_cqc_data.py
@@ -5,7 +5,7 @@ os.environ["SPARK_VERSION"] = "3.3"
 
 from pyspark.sql.dataframe import DataFrame
 
-from jobs.clean_ind_cqc_filled_posts import reduce_dataset_to_earliest_file_per_month
+from utils.cleaning_utils import reduce_dataset_to_earliest_file_per_month
 from utils import utils
 from utils.column_names.ind_cqc_pipeline_columns import (
     PartitionKeys as Keys,
@@ -35,9 +35,9 @@ def main(
     )
     rules = Rules.rules_to_check
 
-    rules[
-        RuleName.size_of_dataset
-    ] = calculate_expected_size_of_cleaned_ind_cqc_dataset(merged_ind_cqc_df)
+    rules[RuleName.size_of_dataset] = (
+        calculate_expected_size_of_cleaned_ind_cqc_dataset(merged_ind_cqc_df)
+    )
 
     check_result_df = validate_dataset(cleaned_ind_cqc_df, rules)
