+++ tests/test_file_schemas.py
@@ -1166,6 +1166,16 @@ class CleaningUtilsSchemas:
             StructField(CQCL.location_id, StringType(), True),
         ]
     )
+    
+    reduce_dataset_to_earliest_file_per_month_schema = StructType(
+        [
+            StructField(CQCLClean.location_id, StringType(), True),
+            StructField(CQCLClean.import_date, StringType(), True),
+            StructField(Keys.year, StringType(), True),
+            StructField(Keys.month, StringType(), True),
+            StructField(Keys.day, StringType(), True),
+        ]
+    )
 
 
 @dataclass
@@ -1459,15 +1469,7 @@ class IndCQCDataUtils:
 
 @dataclass
 class CleanIndCQCData:
-    reduce_dataset_to_earliest_file_per_month_schema = StructType(
-        [
-            StructField(CQCLClean.location_id, StringType(), True),
-            StructField(CQCLClean.import_date, StringType(), True),
-            StructField(Keys.year, StringType(), True),
-            StructField(Keys.month, StringType(), True),
-            StructField(Keys.day, StringType(), True),
-        ]
-    )
+    
 
     merged_schema_for_cleaning_job = StructType(
         [
