+++ utils/cleaning_utils.py
@@ -5,6 +5,8 @@ from pyspark.sql import (
 
 import pyspark.sql.functions as F
 
+from utils.column_names.ind_cqc_pipeline_columns import PartitionKeys as Keys
+
 key: str = "key"
 value: str = "value"
 import_date_s3_uri_format = "yyyyMMdd"
@@ -174,3 +176,22 @@ def add_aligned_date_column(
     )
 
     return primary_df_with_aligned_dates
+
+
+def reduce_dataset_to_earliest_file_per_month(df: DataFrame) -> DataFrame:
+    """
+    Reduce the dataset to the first file of every month.
+
+    This function identifies the date of the first import date in each month and then filters the dataset to those import dates only.
+
+    Args:
+        df (DataFrame): A dataframe containing the partition keys year, month and day.
+
+    Returns:
+        DataFrame: A dataframe with only the first import date of each month.
+    """
+    first_day_in_month = "first_day_in_month"
+    w = Window.partitionBy(Keys.year, Keys.month).orderBy(Keys.day)
+    df = df.withColumn(first_day_in_month, F.first(Keys.day).over(w))
+    df = df.where(df[first_day_in_month] == df[Keys.day]).drop(first_day_in_month)
+    return df
